backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic
email-validator
bcrypt
python-jose[cryptography]
